+++ cmd/setup.go
@@ -19,9 +19,9 @@ import (
 	"github.com/evcc-io/evcc/server"
 	"github.com/evcc-io/evcc/tariff"
 	"github.com/evcc-io/evcc/util"
-	"github.com/evcc-io/evcc/util/community"
 	"github.com/evcc-io/evcc/util/pipe"
 	"github.com/evcc-io/evcc/util/sponsor"
+	"github.com/evcc-io/evcc/util/telemetry"
 	"github.com/samber/lo"
 	"github.com/spf13/viper"
 	"golang.org/x/text/currency"
@@ -52,8 +52,8 @@ func configureEnvironment(conf config) (err error) {
 		err = sponsor.ConfigureSponsorship(conf.SponsorToken)
 	}
 
-	if err == nil && conf.Community {
-		err = community.Create(sponsor.Token)
+	if err == nil && conf.Telemetry {
+		err = telemetry.Create(sponsor.Token)
 	}
 
 	// setup mqtt client listener
